Added CHN channel name property to radio struct, plugin-io and FGCS header.
The purpose is to have the raw client tuned value available, which is nicer
to show e.g. on the statuspage and the mumble comment (this makes it way
more easy to tune to a remote radio frequency with 8.33 channels, since
the 8.33 channel does not correspond to its wave frequency).

This gets written by the new CHN COM UDP field and transmitted to remote
clients together with the radio information and may be used to pretty print
the tuned frequency (mumble comment, statuspage).
The FGCS header was also enhanced, forming the new 1.1 revision
(older revisions cease to work, because we are still in non-production phase).

/!\ This information should not be used for radio wave calculations, because
this may be altered from the concrete radio model in use (that may yield a very
different real wave frequency for tuning).