Change handling of PTT; add support for mumbles internal PTT binding
PTT handling was moved from the UDP-Server to the mumble API callback mumble_onUserTalkingStateChanged().
The UDP COMn_PTT messages now just set a "request" bit in the affected radio. This is then evaluated in
fgcom_handlePTT() to see if the plugin should activate the mumble requestMicrophoneActivationOvewrite()
in order to open the mic.
Wehn the mic opens (either by the override, or by mumbles own ptt binding, or by mumbles voice activation),
the actual radio's ptt-state is calculated and (if changed) broadcast to remote clients.

To know which radios should PTT by mumbles ptt/voiceact, the config file was enhanced:
Configured radios will be set to PTT only when mumbles own activation was used, and not if just the
protocol demand mic-opening by the override in fgcom_handlePTT().
This allows us easily to activate the mapping alongside any active protocol ptt-signals; the user thus
can use both. Because of this, COM1 was activated as default in the config.

(Resolves #110)