FG Addon: Additional ADF radios
Move FG addon nasal code for ADF/RDF logic in a class,
allowing to easily add more ADFs.

Split the RDF logic in two parts:
- A shared parser which reads the RDF signals sent by the plugin as soon
  as they arrive, parse them and send them to the correct ADF radio.
- The slow-running RDF logic, moved to the ADF class, which runs for
  each ADF independently.

The parser is triggered by a listener on the protocol input properties.
With the current protocol design, this is required in order to obtain
all RDF signals: if two signals are transmitted in the same packet,
properties are set to the values of the first signal, then in the same
frame are overwritten by values of the second signal.

Add a second ADF (adf[1]).