Split off Win32 build into separate package
Currently there is a mumble issue with the updater: https://github.com/mumble-voip/mumble/issues/4946
It will break updates on windows when there is more than one *.dll in the ZIP-package.
Therefore the (rarely needed) WIN32 package is split off, the updater is now also aware of this
and try to download a '-win32.zip' with the 32 bit windows build.

Once the mumble bug was fixed, we can revert back to include the win32 binary into the normal package.
(see details in the plugin makefile comment)
For now we need to publish the 32bit zip separately.