Plugin: Rerequest data for known-uninitialized plugin users
If a remote user speaks and is known by the plugin, but the Userdata/location data is not initialized,
it indicates a problem in the data sync process. This theoretically should never happen in normal circumstances.
When it does however, some asking for notifications could have been dropped, so the local client kindly asks
again, which should reestablish comms shortly after receiving the data update.

This is a somewhat workaround fix and no perfect solution, but adds "self heal" capability.

Fix #119