Plugin: Updater now uses header version
Previously the updater did just check the current plugin's version against the tag name.
Now the tag name does'nt matter anymore:
- the most current tag is looked up from the github release page
- then the fgcom-mumble.h header of that tag is fetched from github
- the FGCOM_VERSION_* constants are then parsed

The advantage is that we now can do arbitary releases without the need to always update the plugin.
We can now for example just update the radioGUI or the FGFS addon.