FGFS addon: Add compat mode for when no usable radios are found
Currently we rely on the "operable" roperty being present and deduct
a "standard instrument imlementation" which comes from a <comm-radio>
entry in instrumentation.xml.
However, some old planes have custom (bare minimum) implementations,
where just the "serviceable" prop is set, alongside the frequency.
Such radios cannot be used out of the box.

This commit introduces a dialog popping up when no COM radios are
found, describing the situation and providing a workaround.
The user can click buttons to add COM1 and/or COM2; that adds
a static boolean "operable=1" property to the COMs proptree.
Upon pressing "reload", the addon retries to add radios,
which now succeeds.
If the user opens the FGCom-addon settings dialog, it will also check
for registered radios and reopen the missing radios dialog if its neccessary.
This way, he can reregister in case he accidently pressed "ignore" the first time.
Also, the combar will not open but show an error.

The "good" solution is still to add better systems simulation to
the plane and usage of the standard instruments.

Fix #193