FGFS-Addon: rework handling of legacy /controls/radios/comm-ptt
The default behavior of FGCom-mumble is that each radio has its distinct `ptt` property
to signify that this radio should transmit.

The legacy FGCom had the property `/controls/radios/comm-ptt` and this is still
bound to the default keyboard events (space and shift+space).
Currently the FGCom-mumble XML protocol definition always sends the legacy PTT
field at the end of the udp data stream, and that causes the individual PTT
fields to be ignored.
This commit introduces better handling by not transmitting the legacy field anymore,
but adding a listener to it and translate the legacy value to the individual COM PTT field.

Fix #200