Client/Plugin: Implement optional autojoin of fgcom-mumble channel
This can be set in the ini file.
Set `autoJoinChannel=1` and the plugin wil try to join the first found
fgcom-mumble named channel. Note, that server channel ID defines the order
in which names are recognized (this probably means "channel creation order").

In case the channel is password protected, set `autoJoinChannelPW=<pw>`.
Setting the PW for non protected channels works too, as it will be
ignored (this may change in the future and is depending on mumble client
implementation details).

Fix #189