Server: playback bot cannot shutdown properly when connection lost
The bot tried to shutdown, that is to properly send "i'm done" messages to other clients.
In the case the bot gets disconnected, this results in lua-mumble erros, because no connectoin is
present, so we cannot query channels etc. Sending data without connectoin is pointless too.
So this commit removes this invalid try.

related/continuation of: a6a8b5e2ca70